Enable to switch between model file and model source code

+++ src/common.h
@@ -75,18 +75,18 @@ static RNN_INLINE void rnnoise_free (void *ptr)
 
 /** Copy n elements from src to dst. The 0* term provides compile-time type checking  */
 #ifndef OVERRIDE_RNN_COPY
-#define RNN_COPY(dst, src, n) (memcpy((dst), (src), (n)*sizeof(*(dst)) + 0*((dst)-(src)) ))
+#define RNN_COPY(dst, src, n) (memcpy((dst), (src), (size_t)(n)*sizeof(*(dst)) + (size_t)(0*((dst)-(src))) ))
 #endif
 
 /** Copy n elements from src to dst, allowing overlapping regions. The 0* term
     provides compile-time type checking */
 #ifndef OVERRIDE_RNN_MOVE
-#define RNN_MOVE(dst, src, n) (memmove((dst), (src), (n)*sizeof(*(dst)) + 0*((dst)-(src)) ))
+#define RNN_MOVE(dst, src, n) (memmove((dst), (src), (size_t)(n)*sizeof(*(dst)) + (size_t)(0*((dst)-(src))) ))
 #endif
 
 /** Set n elements of dst to zero */
 #ifndef OVERRIDE_RNN_CLEAR
-#define RNN_CLEAR(dst, n) (memset((dst), 0, (n)*sizeof(*(dst))))
+#define RNN_CLEAR(dst, n) (memset((dst), 0, (size_t)(n)*sizeof(*(dst))))
 #endif
